requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aiosignal==1.3.2
alibabacloud-alidns20150109==3.5.8
alibabacloud-credentials==1.0.2
alibabacloud-credentials-api==1.0.0
alibabacloud-endpoint-util==0.0.3
alibabacloud-gateway-spi==0.0.3
alibabacloud-openapi-util==0.2.2
alibabacloud-tea==0.4.3
alibabacloud-tea-openapi==0.3.15
alibabacloud-tea-util==0.3.13
alibabacloud-tea-xml==0.0.2
aliyun-python-sdk-core==2.16.0
apscheduler==3.11.0
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.3
frozenlist==1.6.2
idna==3.10
jmespath==0.10.0
multidict==6.4.4
propcache==0.3.1
pycparser==2.22
python-dotenv==1.1.0
requests==2.32.3
tzlocal==5.3.1
urllib3==2.4.0
yarl==1.20.0
